+++ src/widgets/explorer/dialoghelper.h
@@ -23,6 +23,7 @@ public:
   Q_INVOKABLE bool confirm(const QString &p_title, const QString &p_text);
   Q_INVOKABLE void notify(const QString &p_title, const QString &p_text);
   Q_INVOKABLE void copyText(const QString &p_text);
+  Q_INVOKABLE QString chooseFile(const QString &p_title, const QString &p_filter);
 
 private:
   QWidget *m_parent = nullptr;